Extract method.
